backupccl: add setting for num split workers, fail scatters to random node
This patch adds a setting to control the parallelism for split and scatters in
generativeSplitAndScatterProcessor, defaulting to 1. This is a workaround
for #96746 as parallel split and scatters sometimes result in a "store error"
that fails the restore.

In addition, for chunks that have failed to scatter, this patch routes the
chunk to a random node instead of the current node. This is necessary as prior
to the generative version, split and scatter processors were on every node,
thus there was no imbalance introduced from routing chunks that have failed to
scatter to the current node. The new generative split and scatter processor is
only on 1 node, and thus would cause the same node to process all chunks that
have failed to scatter.

Release note: None