backupccl: avoid creating spans that start or end on a .Next() key
Remove all calls to key.Next() from generateAndSendImportSpans so that all cover
spans have valid keys as start and end keys.

This fixes an issue where a split can be called on an invalid key that's in the
form of `someValidKey.Next()` during restore. These invalid keys will generally
have a `NULL` at the end of the key, which will result in an error when calling
`EnsureSafeSplits` on this split key. Currently errors from `EnsureSafeSplits`
are ignored, and thus a split will always be attempted on this type of invalid
split key. This split key can land in the middle of a row with column families,
and thus result in failing SQL queries when querying the restored table.

A slight modification was also made to restore span generation to only create
spans out of the start keys of the file instead of both start and end keys.
This is not expected to change much to the generated spans as most of the time
the end key of a file span is the same as the start key of the subsequent file,
but it simplifies the restore span generation logic and may result in slighly
larger spans during restore.

Fixes: #109483

Release note (bug fix): Fixes an issue where a split can be called on an
invalid key that's in the form of someValidKey.Next() during restore. This split
key can land in the middle of a row with column families, and thus result in
failing SQL queries when querying the restored table.